requirements.txt
langchain
langchain-google-genai
python-dotenv
langchain-community
gradi